driver: power: Silence the pointer to int size warning
The pointer returned by of_device_get_match_data() didn't have the
same size as int on 64-bit architectures, causing a compile warning.
Cast the return value to unsigned long and then to int to silence
the warning.
warning: cast to smaller integer type 'int' from 'const void *'
[-Wvoid-pointer-to-int-cast]\n" error, forbidden warning.

+++ drivers/power/supply/qcom/smb1390-charger-psy.c
@@ -2347,7 +2347,7 @@ static int smb1390_probe(struct platform_device *pdev)
 	}
 
 	platform_set_drvdata(pdev, chip);
-	chip->cp_role = (int)of_device_get_match_data(chip->dev);
+	chip->cp_role = (int)(unsigned long)of_device_get_match_data(chip->dev);
 	switch (chip->cp_role) {
 	case CP_MASTER:
 		chip->nchannels = ARRAY_SIZE(cp_master_chans);
